refactor: extract auth-method helper and dedupe logout dialogs
No behavior change, just tidy
